refactor: 헤더 컴포넌트 리팩토링
